Factor out dummy op2 test classes
